requirements.txt
fastapi
uvicorn
kafka-python
numpy
joblib
scikit-learn
google-generativeai
python-dotenv
faker
